fix: 리뷰 평점 쿼리 수정

+++ src/main/kotlin/com/fastcampus/commerce/review/application/response/ProductReviewRatingResponse.kt
@@ -10,7 +10,7 @@ data class ProductReviewRatingResponse(
     companion object {
         fun from(response: ProductReviewRating) =
             ProductReviewRatingResponse(
-                averageRating = response.average.toDouble(),
+                averageRating = response.average,
                 totalCount = response.totalCount,
                 ratingDistribution = RatingDistributionResponse(
                     oneStarCount = response.oneStarCount.toInt(),
